feat: get netlify graph token from headers
