+++ src/lib/secrets.ts
@@ -3,7 +3,7 @@ import { Event as HandlerEvent } from '../function/event'
 import { BaseHandler, HandlerCallback } from '../function/handler'
 import { Response } from '../function/response'
 
-import { getSecrets, HandlerEventWithOneGraph, NetlifySecrets } from './secrets_helper'
+import { getSecrets, NetlifySecrets } from './secrets_helper'
 // Fine-grained control during the preview, less necessary with a more proactive OneGraph solution
 export { getSecrets } from './secrets_helper'
 
@@ -17,12 +17,12 @@ export type HandlerWithSecrets = BaseHandler<Response, ContextWithSecrets>
 export const withSecrets =
   (handler: BaseHandler<Response, ContextWithSecrets>) =>
   async (
-    event: HandlerEventWithOneGraph | HandlerEvent,
+    event: HandlerEvent,
     context: HandlerContext,
     // eslint-disable-next-line promise/prefer-await-to-callbacks
     callback: HandlerCallback<Response>,
   ) => {
-    const secrets = await getSecrets(event as HandlerEventWithOneGraph)
+    const secrets = await getSecrets(event)
 
     return handler(event, { ...context, secrets }, callback)
   }
